1. 优化字段userId->long userId 2. 添加表user_chat_record_t，用于记录用户对应的chat记录

+++ src/main/java/com/speedboot/speedbotagent/dto/UploadFilesDTO.java
@@ -3,11 +3,11 @@ package com.speedboot.speedbotagent.dto;
 import org.springframework.web.multipart.MultipartFile;
 
 public class UploadFilesDTO {
-    int userId;
+    long userId;
 
     MultipartFile[] files;
 
-    public UploadFilesDTO(int userId, MultipartFile[] files) {
+    public UploadFilesDTO(long userId, MultipartFile[] files) {
         this.userId = userId;
         this.files = files;
     }
@@ -15,11 +15,11 @@ public class UploadFilesDTO {
     public UploadFilesDTO() {
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
